Remove NB_UID ENV command
The NB_UID environment variable doesn't need to be set anymore as we are
handling the UID for the running user in the corresponding helm chart now. This
is better because setting the NB_UID in the docker file causes the upstream
image to `chown` the home directory to this uid, but this is slow and gets
slower as the users have more files in their home directory.

We are now making the cluster launch the pod with the running user as 1001 so
this chown is not required