Different approach to speeding up startup
Previously we removed the NB_UID from the environment so that the ./start.sh
script didn't have to CHOWN the home directory. This worked but broke a couple
of things like the ability to SUDO and also pypark's java runtime.

I've added NB_UID back but set another environment variable called CHOWN_HOME
that stops the ./start.sh from having to do the chown on the home directory.