Stop Slow Startup
The `start.sh` runs `usermod -u` before launching jupyter lab. This means that
it will also chown the user's home directory. This makes startup really slow if
softnas is running slowly or is in another AZ from the user's node.

This change makes the usermod happen before the NFS volume is mounted so it will
only need to STAT / CHOWN a small number of files.