allspark: match hadoop-aws/pyspark version
Robin L. is getting yet more strange errors which may be
caused by this mismatch.

Also being more explicit in the `Dockerfile` regarding the
version of `pyspark` and added a comment so that this is
documented.

NOTE: Leaving the version of `com.amazonaws:aws-java-sdk` as
it is for the moment but it's not 100% clear whether this needs
to be specified or `hadoop-aws` pulls/uses the correct version
automatically.